+++ src/mc/alk/arena/util/heroes/Heroes_1_5_2.java
@@ -8,19 +8,19 @@ import org.bukkit.entity.Player;
 public class Heroes_1_5_2 extends HeroesUtil{
 
 	@Override
-	public void setHeroPlayerHealth(Player player, int health) {
+	public void setHeroPlayerHealth(Player player, double health) {
 		PlayerUtil.setHealth(player, health, true);
 	}
 
 	@Override
-	public int getHeroHealth(Player player) {
+	public double getHeroHealth(Player player) {
 		return PlayerUtil.getHealth(player,true);
 	}
 
 	@Override
-	public void setHeroHealthP(Player player, int health) {
-		double val = (double)player.getMaxHealth() * health/100.0;
-		PlayerUtil.setHealth(player, (int) val, true);
+	public void setHeroHealthP(Player player, double health) {
+		double val = player.getMaxHealth() * health/100.0;
+		PlayerUtil.setHealth(player, val, true);
 	}
 
 }
