+++ src/mc/alk/arena/util/heroes/Heroes_pre1_5_2.java
@@ -9,30 +9,30 @@ import com.herocraftonline.heroes.characters.Hero;
 public class Heroes_pre1_5_2 extends HeroesUtil{
 
 	@Override
-	public void setHeroPlayerHealth(Player player, int health) {
+	public void setHeroPlayerHealth(Player player, double health) {
 		Hero hero = getHero(player);
 		if (hero == null){
-			player.setHealth(health);
+			player.setHealth((int)health);
 		} else{
-			hero.setHealth(health);
+			hero.setHealth((int)health);
 			hero.syncHealth();
 		}
 	}
 
 	@Override
-	public int getHeroHealth(Player player) {
+	public double getHeroHealth(Player player) {
 		Hero hero = getHero(player);
 		return hero == null ? player.getHealth() : hero.getHealth();
 	}
 
 	@Override
-	public void setHeroHealthP(Player player, int health) {
+	public void setHeroHealthP(Player player, double health) {
 		Hero hero = getHero(player);
 		if (hero == null){
-			double val = (double)player.getMaxHealth() * health/100.0;
+			double val = player.getMaxHealth() * health/100.0;
 			player.setHealth((int)val);
 		} else{
-			double val = (double)hero.getMaxHealth() * health/100.0;
+			double val = hero.getMaxHealth() * health/100.0;
 			hero.setHealth((int)val);
 			hero.syncHealth();
 		}
